Finally got the basic idea of a suspending actor working
Crearted a proc-macro, which rewrites awaits into suspension points, which enforce dropping for
all references to actor state. This ensures that the implementation of the handle method
is safe, even when there are several of these methods in progress, each awaiting other futures,
even with a reference to actor in scope. Stil needs verification, but I feel good about this one.